requirements.txt

transformers==4.57.1
torch==2.1.0
sentencepiece==0.1.99


pandas==2.0.3
numpy==1.24.0


PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0


flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

tqdm==4.66.1
matplotlib==3.8.0
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
scipy==1.11.4
nltk==3.8.1